Google sign in is enabled
