+++ packages/client/src/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.css';
 import { LandingPage } from './containers/LandingPage/LandingPage.Container';
@@ -6,28 +6,26 @@ import { PageNotFound } from './containers/PageNotFound/PageNotFound.Container';
 import { Navbar } from './components/Navigation/Navbar.component';
 import { Main } from './components/Main/Main';
 import { ToolDetailsPage } from './components/ToolDetailsPage/ToolDetailsPage.component';
-import { AuthContextProvider } from './firebase/AuthContext';
 
 function App() {
+  const [isAuth, setIsAuth] = useState(false);
   return (
     <div className="app">
-      <AuthContextProvider>
-        <Router>
-          {/* properties to imitate logged-in/logged-out state of the navigation */}
-          <Navbar isLogedIn={false} userName="Magdalena" />
-          <Main>
-            <Routes>
-              <Route path="/" element={<LandingPage />} />
-              <Route path="*" element={<PageNotFound />} />
-              <Route path="/tools/:id" element={<ToolDetailsPage />} />
-              <Route path="/about-toolbox" element="" />
-              <Route path="/contact-us" element="" />
-              <Route path="/user-name" element="" />
-              <Route path="/favourites" element="" />
-            </Routes>
-          </Main>
-        </Router>
-      </AuthContextProvider>
+      <Router>
+        {/* properties to imitate logged-in/logged-out state of the navigation */}
+        <Navbar isLogedIn={isAuth} userName="Magdalena" />
+        <Main>
+          <Routes>
+            <Route path="/" element={<LandingPage />} />
+            <Route path="*" element={<PageNotFound />} />
+            <Route path="/tools/:id" element={<ToolDetailsPage />} />
+            <Route path="/about-toolbox" element="" />
+            <Route path="/contact-us" element="" />
+            <Route path="/user-name" element="" />
+            <Route path="/favourites" element="" />
+          </Routes>
+        </Main>
+      </Router>
     </div>
   );
 }
