+++ packages/client/src/components/Navigation/Navbar.component.js
@@ -3,18 +3,14 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import { MobileNavigation } from '../MobileNavigation/MobileNavigation.component';
 import './Navbar.css';
-import { UserAuth } from '../../firebase/AuthContext';
+import { signInWithGoogle } from '../../firebase/firebase';
 
-export const Navbar = ({ isLogedIn, userName, ...props }) => {
+export const Navbar = ({ isAuth, userName, ...props }) => {
   const [isMobileNavOpen, setIsMobileNavOpen] = useState(false);
-  const googleSignIn = UserAuth();
 
-  const handleSignIn = async () => {
-    await googleSignIn();
-  };
   let accountStatus = null;
 
-  if (isLogedIn) {
+  if (isAuth) {
     accountStatus = (
       <div className="log-out-container">
         <div className="user-name-container">
@@ -38,7 +34,11 @@ export const Navbar = ({ isLogedIn, userName, ...props }) => {
   } else {
     accountStatus = (
       <div className="log-in-container">
-        <button type="button" className="navbar-link" onClick={handleSignIn}>
+        <button
+          type="button"
+          className="navbar-link"
+          onClick={signInWithGoogle}
+        >
           <img
             className="navbar-icon"
             src="/assets/vectors/vector-person-not-logged.svg"
@@ -71,7 +71,7 @@ export const Navbar = ({ isLogedIn, userName, ...props }) => {
               <img
                 className="navbar-icon"
                 src={
-                  !isLogedIn
+                  !isAuth
                     ? '/assets/vectors/vector-favourites-empty-heart.svg'
                     : '/assets/vectors/vector-favourites-full-heart.svg'
                 }
@@ -97,6 +97,6 @@ export const Navbar = ({ isLogedIn, userName, ...props }) => {
 };
 
 Navbar.propTypes = {
-  isLogedIn: PropTypes.bool.isRequired,
+  isAuth: PropTypes.bool.isRequired,
   userName: PropTypes.string.isRequired,
 };
