--- packages/client/src/firebase/AuthContext.js
@@ -1,69 +1,0 @@
-import { React, useEffect, useContext, createContext, useState } from 'react';
-import {
-  GoogleAuthProvider,
-  signInWithPopup,
-  signOut,
-  onAuthStateChanged,
-} from 'firebase/auth';
-import { auth } from './firebase';
-import PropTypes from 'prop-types';
-
-const AuthContext = createContext();
-
-export const AuthContextProvider = ({ children }) => {
-  const [user, setUser] = useState({});
-
-  const googleSignIn = () => {
-    const provider = new GoogleAuthProvider();
-    signInWithPopup(auth, provider)
-      .then((result) => {
-        // This gives you a Google Access Token. You can use it to access the Google API.
-        const credential = GoogleAuthProvider.credentialFromResult(result);
-        const token = credential.accessToken;
-        // The signed-in user info.
-        const { userInfo } = result;
-        // ...
-      })
-      .catch((error) => {
-        // Handle Errors here.
-        const errorCode = error.code;
-        const errorMessage = error.message;
-        // The email of the user's account used.
-        const { email } = error.customData;
-        // The AuthCredential type that was used.
-        const credential = GoogleAuthProvider.credentialFromError(error);
-        // ...
-      });
-  };
-
-  const logOut = () => {
-    signOut(auth);
-  };
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-      setUser(currentUser);
-    });
-    return () => {
-      unsubscribe();
-    };
-  }, []);
-
-  return (
-    <AuthContext.Provider value={(googleSignIn, logOut, user)}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const UserAuth = () => {
-  return useContext(AuthContext);
-};
-
-AuthContextProvider.propTypes = {
-  children: PropTypes.element,
-};
-
-AuthContextProvider.defaultProps = {
-  children: PropTypes.element,
-};
